ml model/api/requirements.txt
flask==2.3.3
werkzeug==2.3.7
mysql-connector-python==8.0.26
scikit-learn==1.0.2
numpy==1.21.2
pandas==1.3.3
nltk==3.6.3
flask-cors==3.0.10